Adding branch for ansible to use

+++ ansible.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-ASSETS="CNAME
+ASSETLIST="CNAME
 LICENSE
 README.md
 _config.yml
@@ -17,7 +17,15 @@ index.html
 js
 logout"
 
-for asset in $ASSETS ; do
-    rm -rf $asset
+BRANCHES=`git branch | grep -E '^\*' | cut -d' ' -f 2`
+if [[ $BRANCHES =~ ^ansible$ ]] ; then
+    echo "On Branch ansible"
+    for asset in $ASSETLIST ; do
+        echo $asset
+        rm -rf $asset
+    done
     mv _site site
-done
+else
+    echo "Not on ansible branch"
+    exit 1
+fi
